pyproject.toml
[project]
name = "trip-agent"
version = "0.1.0"
description = "Multi-agent travel planning system with A2A and MCP"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "mcp[cli]>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "mcp_servers", "shared"]
